switch to flint's binary matrices

+++ src/utils.h
@@ -1,6 +1,6 @@
 #ifndef UTILS_H
 #define UTILS_H
 
-void generate_random_set(int upper_bound, int size, int set[size]);
+void generate_random_set(unsigned long upper_bound, unsigned long size, unsigned long set[size]);
 
 #endif
